+++ src/composables/scenarioExport.ts
@@ -1,10 +1,11 @@
-import { feature, featureCollection, point } from "@turf/helpers";
+import { featureCollection, point } from "@turf/helpers";
 import { injectStrict } from "@/utils";
 import { activeScenarioKey } from "@/components/injects";
 import { TScenario } from "@/scenariostore";
 import { ExportSettings } from "@/types/convert";
 import * as FileSaver from "file-saver";
 import { symbolGenerator } from "@/symbology/milsymbwrapper";
+import type { Root } from "@tmcw/togeojson";
 
 export interface UseScenarioExportOptions {
   activeScenario: TScenario;
@@ -44,17 +45,26 @@ export function useScenarioExport(options: Partial<UseScenarioExportOptions> = {
     );
   }
 
-  async function createKMLString(opts: ExportSettings) {
-    const { foldersToKML } = await import("@placemarkio/tokml");
+  async function createKMLString(sidcs: string[], opts: ExportSettings) {
+    const { foldersToKML } = await import("@/extlib/tokml");
     const units = opts.includeUnits ? convertUnitsToGeoJson().features : [];
     const features = opts.includeFeatures
       ? convertScenarioFeaturesToGeoJson().features
       : [];
 
-    const root = { type: "root", children: [] as any[] };
+    const root: Root = { type: "root", children: [] };
 
     if (opts.includeUnits) {
-      root.children.push({ type: "folder", meta: { name: "Units" }, children: units });
+      root.children.push({
+        type: "folder",
+        meta: { name: "Units" },
+        children: units.map((unit) => {
+          return {
+            ...unit,
+            properties: { ...unit.properties, styleUrl: `#sidc${unit.properties.sidc}` },
+          };
+        }),
+      });
     }
 
     if (opts.includeFeatures) {
@@ -64,11 +74,11 @@ export function useScenarioExport(options: Partial<UseScenarioExportOptions> = {
         children: features,
       });
     }
-    return foldersToKML(root);
+    return foldersToKML(root, sidcs);
   }
 
   async function downloadAsKML(opts: ExportSettings) {
-    const kmlString = await createKMLString(opts);
+    const kmlString = await createKMLString([], opts);
     FileSaver.saveAs(
       new Blob([kmlString], {
         type: "application/vnd.google-earth.kml+xml",
@@ -78,16 +88,15 @@ export function useScenarioExport(options: Partial<UseScenarioExportOptions> = {
   }
 
   async function downloadAsKMZ(opts: ExportSettings) {
-    const kmlString = await createKMLString(opts);
     const { zipSync } = await import("fflate");
-
     const data: Record<string, Uint8Array> = {};
-    data["doc.kml"] = new TextEncoder().encode(kmlString);
+    const usedSidcs = new Set<string>();
     if (opts.embedIcons) {
       for (const unit of geo.everyVisibleUnit.value) {
         const { sidc } = unit;
-        if (!(sidc in data)) {
+        if (!usedSidcs.has(sidc)) {
           const symb = symbolGenerator(sidc);
+          usedSidcs.add(sidc);
           const blob: Blob | null = await new Promise((resolve) =>
             symb.asCanvas().toBlob(resolve)
           );
@@ -97,6 +106,9 @@ export function useScenarioExport(options: Partial<UseScenarioExportOptions> = {
         }
       }
     }
+    const kmlString = await createKMLString([...usedSidcs], opts);
+
+    data["doc.kml"] = new TextEncoder().encode(kmlString);
 
     const zipData = zipSync(data);
     FileSaver.saveAs(